+++ src/main/kotlin/Application.kt
@@ -1,10 +1,18 @@
 package com.dlfsystems
 
-import com.dlfsystems.app.configureRouting
+import com.dlfsystems.server.Websocket
 import com.dlfsystems.server.Yegg
+import io.ktor.serialization.kotlinx.json.*
 import io.ktor.server.application.*
 import io.ktor.server.engine.*
 import io.ktor.server.netty.*
+import io.ktor.server.plugins.contentnegotiation.*
+import io.ktor.server.request.*
+import io.ktor.server.response.*
+import io.ktor.server.routing.*
+import io.ktor.server.websocket.*
+import io.ktor.websocket.*
+import kotlin.time.Duration.Companion.seconds
 
 fun main() {
     embeddedServer(Netty, port = 8080, host = "0.0.0.0", module = Application::module)
@@ -14,6 +22,33 @@ fun main() {
 fun Application.module() {
 
     Yegg.start()
-    configureRouting()
+
+    install(ContentNegotiation) {
+        json()
+    }
+
+    install(WebSockets) {
+        pingPeriod = 15.seconds
+        timeout = 15.seconds
+        maxFrameSize = Long.MAX_VALUE
+        masking = false
+    }
+
+    routing {
+        webSocket("/ws") {
+            Websocket(incoming, outgoing) { close(it) }.listen()
+        }
+
+        post("/program/{traitName}/{verbName}")  {
+            val code = call.receiveText()
+            var result = ""
+            call.parameters["traitName"]?.also { traitName ->
+                call.parameters["verbName"]?.also { verbName ->
+                    result = Yegg.world.programVerb(traitName, verbName, code)
+                }
+            }
+            call.respondText(result)
+        }
+    }
 
 }
