+++ src/main/kotlin/server/Yegg.kt
@@ -1,7 +1,7 @@
 package com.dlfsystems.server
 
-import com.dlfsystems.app.Log
 import com.dlfsystems.server.mcp.MCP
+import com.dlfsystems.server.parser.Connection
 import com.dlfsystems.value.*
 import com.dlfsystems.world.World
 import com.dlfsystems.world.Obj
