+++ src/main/kotlin/server/parser/Connection.kt
@@ -1,10 +1,9 @@
-package com.dlfsystems.server
+package com.dlfsystems.server.parser
 
-import com.dlfsystems.app.Log
+import com.dlfsystems.server.Log
+import com.dlfsystems.server.Yegg
 import com.dlfsystems.server.mcp.MCP
 import com.dlfsystems.server.mcp.Task
-import com.dlfsystems.server.parser.CommandMatch
-import com.dlfsystems.server.parser.Preposition
 import com.dlfsystems.util.NanoID
 import com.dlfsystems.world.Obj
 import com.dlfsystems.world.trait.Verb
